h3lis100dl: fix acceleration_raw_get() API
acceleration_raw_get() API must read only 0x29/0x2B/0x2D
regiters (0x28/0x2A/0x2C are reserved).
(fix issue #1)

+++ h3lis100dl_reg.h
@@ -537,7 +537,7 @@ int32_t h3lis100dl_flag_data_ready_get(stmdev_ctx_t *ctx,
                                        uint8_t *val);
 
 int32_t h3lis100dl_acceleration_raw_get(stmdev_ctx_t *ctx,
-                                        int16_t *val);
+                                        int8_t *val);
 
 int32_t h3lis100dl_device_id_get(stmdev_ctx_t *ctx, uint8_t *buff);
